[JIT] Modify exit transform pass to unify types to Any
**Summary**
This commit modifies the type unification performed during the exit
transform pass to add if node outputs so that it falls back to Any
instead of nothing (c10::nullopt). If type unification fails in this
situation, it means that the two branches of the if return different
types, so the output if the if node should be of type Any.

Issue #41962 reported that the use of an `Any` return type in
combination with different code paths returning values of different
types causes a segmentation fault. This is because the exit transform
pass tries to unify the different return types, fails, but silently sets
the type of the if node to c10::nullopt. This causes problems later in
shape analysis when that type object is dereferenced.

**Test Plan**
This commit adds a unit test that checks that a function similar to the
one in #41962 can be scripted and executed.

**Fixes**
This commit fixes #41962.

ghstack-source-id: 84adccd91a08c4eddbbadc81b861d8df06b324dc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42259